+++ src/webgis4u/ol/control.js
@@ -2,9 +2,10 @@
  * @module webgis4u/ol/control
  */
 
-export { default as PanBar } from './control/PanBar';
 export { default as LayerCheckbox } from './control/LayerCheckbox';
 export { default as LayerRadioButton } from './control/LayerRadioButton';
 export { default as LegendDisplayable } from './control/LegendDisplayable';
 export { default as LegendVisibility } from './control/LegendVisibility';
+export { default as Measure } from './control/Measure';
 export { default as OverviewMap } from './control/OverviewMap';
+export { default as PanBar } from './control/PanBar';